// src/pokemons/types/pokemon-type.ts
export enum PokemonType {
  Normal = 'Normal',
  Fire = 'Fire',
  Water = 'Water',
  Grass = 'Grass',
  Rock = 'Rock',
  Flying = 'Flying',
  Electric = 'Electric',
  Ground = 'Ground',
  Fairy = 'Fairy',
  Fighting = 'Fighting',
  Ice = 'Ice',
  Psychic = 'Psychic',
  Bug = 'Bug',
  Ghost = 'Ghost',
  Dark = 'Dark',
  Dragon = 'Dragon',
  Poison = 'Poison',
  Steel = 'Steel',
}